extract test for Google Proxy
