+++ static/src/javascripts/lib/detect-google-proxy.ts
@@ -1,0 +1,5 @@
+export const isGoogleProxy = (): boolean =>
+	Boolean(
+		navigator.userAgent.includes('Google Web Preview') ||
+			navigator.userAgent.includes('googleweblight'),
+	);
